Adding tempest tests for deployment and update of multiple policies and rules
Issues:
Fixes #324

Problem:
New tests are needed to validate whether the agent does the correct
thing when policies with multiple rules are deleted or when rules
are deleted in varying orders. Also need to added validation that the
proper conditions exist on the device when a rule is deployed. This is just
a test effort.

Analysis:
Added a couple of new tests and changed the validation of all api
tests to check the value of a condition in addition to the comparison
type.  This means checking that header is equal to 'test', instead of just
checking whether a condition exists that checks a header. This
allows us to distinguish between different header conditions in the same rule.

Tests:
All api and scenario tests pass