Newton process creation event changed from mitaka
Issues:
Fixes #528

Problem:
When a neutron worker is created in mitaka, the AFTER_CREATE event is
emitted, yet when the worker is created in newton, the AFTER_INIT event
is emitted. Now that we wait for these processes to be forked before we
create the RPC listener in the driver, we must rely upon the proper
event before agent and driver communication can commence.

Analysis:
Changed the AFTER_CREATE event to AFTER_INIT and communication then
flowed between driver and agent.

Tests:
Ran manual tests for creation of loadbalancer, then automated testing
with all tests.